Move VectorUtils into Maxsat Backend package.
